Tweak thumbnail sizes a bit so it the NNTP looks better side by side on a 1920x1200 screen.
Also, tweak the default size of new windows slightly. In case we decide
to use side by side windows size the window in such a way that 2
windows fit on the screen and have the kWindowTilePixels spacing
between each other and the screen edges.

BUG=None

TEST=Show two windows side. On a 1920x1200 screen they should both be
showing the thumbnails in normal size.

Review URL: http://codereview.chromium.org/155930

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21364 0039d316-1c4b-4281-b951-d872f2087c98