Add SetRect to Texture2d and TextureCUBE
Originally I was going to replace Lock/Unlock with
only SetRect (and maybe GetRect) and I did that.

But then as I was going through the code things started
not being so cool. The JavaScript texture.SetRect where 
as before it allocated no memory, now it needed to allocate
a temporary buffer just to be able to write the texture.

DrawImage also would have have required copying the original 
texture out using GetRect, then modifying it, then copying it 
back even in D3D or it would have required a temporary buffer.

That's what happens under the hood in the GL implementation
of Texture2D but why should D3D suffer because of the of GL?
Or maybe it's a reasonable compomise the D3D should go slower
so that GL isn't twice as slow as it it could be? 

So, I left Lock/Unlock and added SetRect as well.

This CL should help the video guys and we can decide if 
we want to get rid of Lock/Unlock later.

SetRect is really only there because trying to make GL
act like D3D makes GL slow since we had to get a copy
of the texture from GL before Locking.  SetRect, since
it is not exposing the original data doesn't need to do
that.

So, now there are both. If need to read (and write) to
the texture use Lock.  If you only need to write use
SetRect.

I also fixed Lock/Unlock so they return a pitch which
is required for D3D. While I did that I made Lock/Unlock
private so you have to use the TextureXXXLockHelpers
to access the data.

I didn't do the command buffer versions.

I also added stubs for texture tests. We need to add
more tests.

I feel like a lot of clean up needs to happen.

It seesm like Bitmap should be split into

Bitmap2D
BitmapCUBE
BitmapVolume

Possibly. And maybe BitmapCube should effectively
just be typedef Bitmap2D BitmapCUBE[6] and 
BitmapVolume would be std::vector<Bitmap2D>

Then we wouldn't need face and volume versions
of various functions. You'd just get the face or
the slice Bitma2D and then use the 2D functions.

We should decide if that's what we want before
pushing a new release. Otherwise we should remove
bitmap.idl from idl_scons.lst and the one sample

And ship without exposing Bitmap until we are
sure they API we are exposing is the one we want.

Review URL: http://codereview.chromium.org/159725

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@22332 0039d316-1c4b-4281-b951-d872f2087c98