Run the DOM Storage layout tests in the UI test framework since the test shell doesn't hit very much of the new code.  I still plan to write unit tests and we're still goign to enable the layout tests in test shell, but this is still a good way for us to test the full stack ASAP.
Workers has used this technique for a while now.  This CL factors out the layout test running code from the workers ui test and moves arounds directories a bit so not everything is in the "workers" directory.

Doing this via a v8 extension is the right way to go long term, but I ran into a couple snags, so I think this is a good first step.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/159720

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@22466 0039d316-1c4b-4281-b951-d872f2087c98