Reland r23358 with a fix for linux unit_tests:
Better location for setting the size of tab contents.  This should
catch all code paths.  I noticed that middle clicking a bookmark
wasn't working, so I found a single place to do this.

This code is always run when a tab is added and it sizes the
contents when it's loading in the background.

BUG=619

TBR=japhet

Review URL: http://codereview.chromium.org/164526


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23384 0039d316-1c4b-4281-b951-d872f2087c98