A quick fix for Issue 18844 (Take 2).
This is the same change as <http://codereview.chromium.org/164309> that I reverted to investigate a test_shell_test failure (Issue 19263).

This change checks if the modifier flag is a WebInputEvent::MetaKey flag to
switch the scroll granularity when a user types a up/down key.
Editor::Commands ("MoveToBeginningOfDocument" and "MoveToEndOfDocument") works
only in an editable control, such as <textarea>. On the other hand, this
WebViewImpl::ScrollViewWithKeyboard() is for scrolling a page.

BUG=18844 "Apple (command) key + down arrow is not working"
TEST=Type command+down keys in a web page which contains a scroll bar, and
verify Chrome scrolls the page to its end.
Review URL: http://codereview.chromium.org/165523

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23540 0039d316-1c4b-4281-b951-d872f2087c98