Move WebPluginContainer into the WebKit API
This change introduces the following new interfaces:

  WebPlugin
  WebPluginContainer

The implementation of WebPluginContainer is moved to webkit/api/src/WebPluginContainerImpl.{h,cpp}, and WebPluginImpl from webkit/glue now implements WebKit::WebPlugin.

This change does not completely eliminate dependencies in WebPluginImpl on WebCore.  The rest will be removed in a separate change.

A hack was added to TemporaryGlue to deal with the fact that ChromeClientImpl still resides in webkit/glue.  Also, StackingOrderIterator in webkit/glue because there is another CL out to remove that from the source tree.

Note: This change depends on another CL that I am brewing to make WebCanvas be a CGContext on Mac.

R=jam
BUG=10036
TEST=covered by existing tests
Review URL: http://codereview.chromium.org/165464

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23797 0039d316-1c4b-4281-b951-d872f2087c98