Fix document.hasFocus() for layout test
We were not setting the focused frame to null when we blurred a window.
Therefore a document inside a frame inside a blurred window would think
it was focused since its frame was the focused frame.

This fixes the layout test for the test_shell but Chrome is still
having issues with the test since blurring is asynchronous in chrome.
Changing the layout test by adding timers makes the test pass in chrome
as well.

BUG=16018

TEST=Run LayoutTests/fast/dom/HTMLDocument/hasFocus.html

Review URL: http://codereview.chromium.org/182032

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@24947 0039d316-1c4b-4281-b951-d872f2087c98