ForceTLS: generate internal redirect when mistakenly using http.
Previously, we just generated a cryptic error when the user attempted
to navigate to a ForceTLS site with HTTP. With this patch, we generate
an internal redirect to the HTTPS version of the site.

(This collided with Dave's r25496, which also added
url_request_redirect_job.cc. Thankfully our code was almost identical.)

http://codereview.chromium.org/193014


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@25519 0039d316-1c4b-4281-b951-d872f2087c98