GTK Theme: Don't render IDR_THEME_TOOLBAR on GtkWidgets.
IDR_THEME_TOOLBAR is a generated blank image of the bg[SELECTED] color,
which will be drawn by default on GtkWidgets. Uploading an image of
said color to the X server on each expose and then asking it to tile the
image across the widget is silly; stop doing it.

Also sets clip rectangles correctly in a few places where we were using GdkGC
instead of skia.

Review URL: http://codereview.chromium.org/196044

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@25647 0039d316-1c4b-4281-b951-d872f2087c98