Expectations for 3 perf systems on morejs and startup.
Set the perf expectations on linux, mac, and windows
perf systems for the morejs and warm startup tests.
The variance assumptions might require fine-tuning on
the master to get right.  We might be flaky to begin
with, or may want to turn off the variance reporting
altogether.

BUG=18597
TEST=build.chromium.org perf slaves stay green
Review URL: http://codereview.chromium.org/193083

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26376 0039d316-1c4b-4281-b951-d872f2087c98