NTP: Fix startup visual state
This inject the template data earlier in the document so that we can set
the initial visual state based on the user preference. This prevents
flickering when the user is not showing the most visited thumbnails.

BUG=24513

TEST=Use a theme with a background image to make this easier to spot.
Hide the most visited. Restart Chrome. The thumbnails should not flicker

Review URL: http://codereview.chromium.org/269095

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@29314 0039d316-1c4b-4281-b951-d872f2087c98